changes for common component for ask for help and others
